Split ui in multiple modules
