+++ src/ui/header.rs
@@ -1,0 +1,22 @@
+use tui::{
+    backend::Backend,
+    layout::Rect,
+    style::Style,
+    text::{Span, Spans},
+    widgets::{Block, Borders, Paragraph, Wrap},
+    Frame,
+};
+
+pub fn draw<B>(f: &mut Frame<B>, area: Rect)
+where
+    B: Backend,
+{
+    let text = vec![Spans::from(
+        "This will be the section explaining the graham scan algorithm.",
+    )];
+    let block = Block::default()
+        .borders(Borders::ALL)
+        .title(Span::styled("Description", Style::default()));
+    let paragraph = Paragraph::new(text).block(block).wrap(Wrap { trim: true });
+    f.render_widget(paragraph, area);
+}
